# web/openapi.json
{
  "components": {
    "schemas": {
      "AnomaliaSalida": {
        "properties": {
          "creada_en": {
            "format": "date-time",
            "title": "Creada En",
            "type": "string"
          },
          "despacho_id_externo": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Despacho Id Externo"
          },
          "detalle": {
            "additionalProperties": true,
            "title": "Detalle",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "resuelta": {
            "title": "Resuelta",
            "type": "boolean"
          },
          "severidad": {
            "enum": [
              "ALTA",
              "AVISO",
              "INFO"
            ],
            "title": "Severidad",
            "type": "string"
          },
          "surtidor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surtidor Id"
          },
          "tipo": {
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo",
          "severidad",
          "surtidor_id",
          "despacho_id_externo",
          "detalle",
          "creada_en",
          "resuelta"
        ],
        "title": "AnomaliaSalida",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "detail": {
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "FilaTurnoSalida": {
        "properties": {
          "despachos": {
            "title": "Despachos",
            "type": "integer"
          },
          "estado": {
            "enum": [
              "ABIERTO",
              "CERRADO"
            ],
            "title": "Estado",
            "type": "string"
          },
          "fin": {
            "format": "date-time",
            "title": "Fin",
            "type": "string"
          },
          "galones": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Galones",
            "type": "string"
          },
          "id_cierre": {
            "title": "Id Cierre",
            "type": "integer"
          },
          "inicio": {
            "format": "date-time",
            "title": "Inicio",
            "type": "string"
          },
          "lado": {
            "title": "Lado",
            "type": "string"
          },
          "lectura_final_valor": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lectura Final Valor"
          },
          "lectura_final_vol": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lectura Final Vol"
          },
          "surtidor_id": {
            "title": "Surtidor Id",
            "type": "integer"
          },
          "turno_id": {
            "title": "Turno Id",
            "type": "integer"
          },
          "valor": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Valor",
            "type": "string"
          }
        },
        "required": [
          "turno_id",
          "surtidor_id",
          "id_cierre",
          "estado",
          "inicio",
          "fin",
          "lado",
          "despachos",
          "galones",
          "valor",
          "lectura_final_vol",
          "lectura_final_valor"
        ],
        "title": "FilaTurnoSalida",
        "type": "object"
      },
      "FilaVentasSalida": {
        "properties": {
          "despachos": {
            "title": "Despachos",
            "type": "integer"
          },
          "forma_pago": {
            "enum": [
              "CONTADO",
              "CREDITO",
              "OTRO"
            ],
            "title": "Forma Pago",
            "type": "string"
          },
          "galones": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Galones",
            "type": "string"
          },
          "origen": {
            "description": "MANUAL: digitada en la pista, sin despacho enlazado",
            "enum": [
              "SURTIDOR",
              "MANUAL"
            ],
            "title": "Origen",
            "type": "string"
          },
          "periodo": {
            "format": "date",
            "title": "Periodo",
            "type": "string"
          },
          "producto": {
            "title": "Producto",
            "type": "string"
          },
          "surtidor_id": {
            "title": "Surtidor Id",
            "type": "integer"
          },
          "valor": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Valor",
            "type": "string"
          }
        },
        "required": [
          "surtidor_id",
          "periodo",
          "producto",
          "forma_pago",
          "origen",
          "despachos",
          "galones",
          "valor"
        ],
        "title": "FilaVentasSalida",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PosibleDuplicadoSalida": {
        "properties": {
          "despacho_id": {
            "title": "Despacho Id",
            "type": "integer"
          },
          "galones": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Galones",
            "type": "string"
          },
          "inicio_despacho": {
            "format": "date-time",
            "title": "Inicio Despacho",
            "type": "string"
          },
          "lado": {
            "title": "Lado",
            "type": "string"
          },
          "producto": {
            "title": "Producto",
            "type": "string"
          },
          "registrada_en": {
            "format": "date-time",
            "title": "Registrada En",
            "type": "string"
          },
          "surtidor_id": {
            "title": "Surtidor Id",
            "type": "integer"
          },
          "valor": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Valor",
            "type": "string"
          },
          "venta_manual_id": {
            "title": "Venta Manual Id",
            "type": "integer"
          }
        },
        "required": [
          "venta_manual_id",
          "despacho_id",
          "surtidor_id",
          "lado",
          "producto",
          "galones",
          "valor",
          "registrada_en",
          "inicio_despacho"
        ],
        "title": "PosibleDuplicadoSalida",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "token de POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema que reemplaza a Nexus. **Banco de pruebas: no envía nada a la DIAN.** Dinero y galones viajan como texto decimal.",
    "title": "Estación de servicio — banco de pruebas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/anomalias": {
      "get": {
        "operationId": "listar_anomalias_anomalias_get",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "default": "abiertas",
              "enum": [
                "abiertas",
                "resueltas",
                "todas"
              ],
              "title": "Estado",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "severidad",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ALTA",
                    "AVISO",
                    "INFO"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severidad"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnomaliaSalida"
                  },
                  "title": "Response Listar Anomalias Anomalias Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Anomalías para revisión humana",
        "tags": [
          "reportes"
        ]
      }
    },
...
    "/reportes/contador": {
      "get": {
        "operationId": "contador_reportes_contador_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "required": false,
            "schema": {
              "default": "xlsx",
              "enum": [
                "xlsx",
                "csv",
                "pdf"
              ],
              "title": "Formato",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {},
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {},
              "text/csv": {}
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ventas y facturas para la contabilidad (formato PROVISIONAL; lleva datos personales)",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/diario": {
      "get": {
        "operationId": "diario_reportes_diario_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "surtidor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Surtidor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FilaVentasSalida"
                  },
                  "title": "Response Diario Reportes Diario Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ventas por día, producto y forma de pago",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/exportar": {
      "get": {
        "operationId": "exportar_reportes_exportar_get",
        "parameters": [
          {
            "in": "query",
            "name": "surtidor_id",
            "required": true,
            "schema": {
              "title": "Surtidor Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Excel de ventas de un surtidor (Diario, Mensual, Turnos, Datos, Alertas, Notas)",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/mensual": {
      "get": {
        "operationId": "mensual_reportes_mensual_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "surtidor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Surtidor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FilaVentasSalida"
                  },
                  "title": "Response Mensual Reportes Mensual Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ventas por mes, producto y forma de pago",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/pdf": {
      "get": {
        "operationId": "reporte_pdf_reportes_pdf_get",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": true,
            "schema": {
              "enum": [
                "diario",
                "mensual",
                "turno"
              ],
              "title": "Tipo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "surtidor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Surtidor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reporte diario, mensual o por turno en PDF (sin datos personales)",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/posibles-duplicados": {
      "get": {
        "operationId": "posibles_duplicados_reportes_posibles_duplicados_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PosibleDuplicadoSalida"
                  },
                  "title": "Response Posibles Duplicados Reportes Posibles Duplicados Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ventas manuales sin enlace que parecen ser un despacho ya contado (posible doble conteo)",
        "tags": [
          "reportes"
        ]
      }
    },
    "/reportes/turno": {
      "get": {
        "operationId": "turno_reportes_turno_get",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Desde",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Hasta",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "surtidor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Surtidor Id"
            }
          },
          {
            "in": "query",
            "name": "turno_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Turno Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FilaTurnoSalida"
                  },
                  "title": "Response Turno Reportes Turno Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resumen por turno (ID-CIERRE) y lado",
        "tags": [
          "reportes"
        ]
      }
    },
...
  }
}